Refactor into separate Appointment component.
